changed int to double

+++ Calculator/src/Calculator.java
@@ -5,10 +5,10 @@ public class Calculator {
 		
 		Scanner sc = new Scanner(System.in);
 		System.out.println("Input the first Number: ");
-		int x = sc.nextInt();
+		double x = sc.nextDouble;
 		
 		System.out.println("Input the second number: ");
-		int y = sc.nextInt();
+		double y = sc.nextDouble();
 		
 		System.out.println("The addition of the two numbers are: " + addition(x,y));
 		
@@ -17,7 +17,7 @@ public class Calculator {
 		
 	}
 
-	public static int addition(int a, int b) {
+	public static double addition(Double a, Double b) {
 		
 		return a+b;
 	}
